+++ src/_packages/core.index.ts
@@ -3,7 +3,7 @@
 import { prettifyPromptbookString } from '../conversion/prettify/prettifyPromptbookString';
 import { promptbookJsonToString } from '../conversion/promptbookJsonToString';
 import { promptbookStringToJson } from '../conversion/promptbookStringToJson';
-import { validatePromptbookJson } from '../conversion/validation/validatePromptbookJson';
+import { validatePromptbook } from '../conversion/validation/validatePromptbook';
 import { ExpectError } from '../errors/_ExpectError';
 import { PromptbookExecutionError } from '../errors/PromptbookExecutionError';
 import { PromptbookLibraryError } from '../errors/PromptbookLibraryError';
@@ -61,7 +61,7 @@ export { createLibraryFromJson, createLibraryFromPromise, createLibraryFromUrl, 
 export { SimplePromptInterfaceTools };
 
 // @promptbook/parser
-export { promptbookJsonToString, promptbookStringToJson, validatePromptbookJson };
+export { promptbookJsonToString, promptbookStringToJson, validatePromptbook };
 
 // @promptbook/executor
 export { createPromptbookExecutor, MultipleLlmExecutionTools };