+++ src/conversion/validation/validatePromptbook-logicErrors.test.ts
@@ -1,14 +1,14 @@
 import { describe, expect, it } from '@jest/globals';
 import { promptbookStringToJson } from '../promptbookStringToJson';
 import { importPromptbook } from './_importPromptbook';
-import { validatePromptbookJson } from './validatePromptbookJson';
+import { validatePromptbook } from './validatePromptbook';
 
-describe('validatePromptbookJson', () => {
+describe('validatePromptbook', () => {
     it('should fail on using parameter that is not defined', () => {
         expect(async () => {
             const promptbookString = importPromptbook('errors/logic/undefined-parameter.ptbk.md');
             const promptbookJson = await promptbookStringToJson(promptbookString);
-            validatePromptbookJson(promptbookJson);
+            validatePromptbook(promptbookJson);
         }).rejects.toThrowError(/Can not resolve some parameters/i);
     });
 
@@ -16,7 +16,7 @@ describe('validatePromptbookJson', () => {
         expect(async () => {
             const promptbookString = importPromptbook('errors/logic/unused-parameter.ptbk.md');
             const promptbookJson = await promptbookStringToJson(promptbookString);
-            validatePromptbookJson(promptbookJson);
+            validatePromptbook(promptbookJson);
         }).rejects.toThrowError(/Parameter \{name\} is created but not used/i);
     });
 
@@ -24,7 +24,7 @@ describe('validatePromptbookJson', () => {
         expect(async () => {
             const promptbookString = importPromptbook('errors/logic/model-mismatch.ptbk.md');
             const promptbookJson = await promptbookStringToJson(promptbookString);
-            validatePromptbookJson(promptbookJson);
+            validatePromptbook(promptbookJson);
         }).rejects.toThrowError(/Unknown model key/i);
     });
 
@@ -32,7 +32,7 @@ describe('validatePromptbookJson', () => {
         expect(async () => {
             const promptbookString = importPromptbook('errors/logic/wrong-expectations.ptbk.md');
             const promptbookJson = await promptbookStringToJson(promptbookString);
-            validatePromptbookJson(promptbookJson);
+            validatePromptbook(promptbookJson);
         }).rejects.toThrowError(/Max expectation of words must be positive/i);
     });
 
@@ -40,7 +40,7 @@ describe('validatePromptbookJson', () => {
         expect(async () => {
             const promptbookString = importPromptbook('errors/logic/joker-without-expectations.ptbk.md');
             const promptbookJson = await promptbookStringToJson(promptbookString);
-            validatePromptbookJson(promptbookJson);
+            validatePromptbook(promptbookJson);
         }).rejects.toThrowError(/Joker parameters are used for \{name\} but no expectations are defined/i);
     });
 
@@ -48,13 +48,13 @@ describe('validatePromptbookJson', () => {
         expect(async () => {
             const promptbookString = importPromptbook('errors/logic/circular-parameters-simple.ptbk.md');
             const promptbookJson = await promptbookStringToJson(promptbookString);
-            validatePromptbookJson(promptbookJson);
+            validatePromptbook(promptbookJson);
         }).rejects.toThrowError(/circular dependencies/i);
 
         expect(async () => {
             const promptbookString = importPromptbook('errors/logic/circular-parameters-advanced.ptbk.md');
             const promptbookJson = await promptbookStringToJson(promptbookString);
-            validatePromptbookJson(promptbookJson);
+            validatePromptbook(promptbookJson);
         }).rejects.toThrowError(/circular dependencies/i);
     });
 });