+++ src/library/constructors/createLibraryFromJson.test.ts
@@ -34,7 +34,7 @@ describe('createLibraryFromJson', () => {
     it('should get promptbook by url from library', async () => {
         expect.assertions(1);
         const promptbook = await promptbookStringToJson(promptbookString);
-        const library = await createLibraryFromJson(promptbook);
+        const library = createLibraryFromJson(promptbook);
         const promptbookFromLibrary = await library.getPromptbookByUrl('https://example.com/promptbook.json');
         expect(promptbookFromLibrary).toEqual(await promptbookStringToJson(promptbookString));
     });