+++ CHANGELOG.md
@@ -347,7 +347,7 @@ System for management of external knowledge (RAG)
 -   Add `KnowledgeJson`
 -   Add `prepareKnowledgeFromMarkdown` exported from `@promptbook/core`
 -   Change `promptbookStringToJson` to async function
--   Change `createPromptbookLibraryFromSources` to `createPromptbookLibraryFromJson` and allow only compiled jsons as input
+-   Change `createPromptbookLibraryFromSources` to `createPromptbookLibraryFromJson` and allow only compiled jsons as input + it is not `async` anymore
 -   Allow only jsons as input in `createLibraryFromPromise`
 -   Class `SimplePromptbookLibrary` not exported anymore directly, only as type
 -   Rename all `createPromptbookLibraryFromXyz` to `createLibraryFromXyz`
@@ -355,6 +355,7 @@ System for management of external knowledge (RAG)
 -   Add util `libraryToJson` exported from `@promptbook/core`
 -   CLI util `promptbook make ...` can convert promptbooks to JSON
 -   `promptbookStringToJson` automatically looks for `promptbook-library.json` in root of given directory
+-   Rename `validatePromptbookJson` to `validatePromptbook`
 
 ## In prepare